% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Function to plot survival curve}
\usage{
plot_survival(
  dbms = c("oracle", "postgresql", "redshift", "sql server", "pdw", "netezza",
    "bigquery", "sqlite"),
  user,
  password,
  server,
  port = NULL,
  schema = NULL
)
}
\arguments{
\item{dbms}{DBMS runnong on the OMOP server}

\item{survival_time_col}{numberic variable that indicates survival time}

\item{event_col}{boolean variable that indicates event occurrence}

\item{cohort_cols}{variable names of length 1 or greater that indicates grouping variables for output}

\item{pval}{TRUE if pval is to be included in the plot}

\item{median_survival_time}{TRUE if median survival time should be returned}
}
\description{
Function to plot survival curve
}
