% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_to_rx_hist.R
\name{plot_time_to_rx_hist}
\alias{plot_time_to_rx_hist}
\title{Plot a histogram of the time to treatment}
\usage{
plot_time_to_rx_hist(
  dbms = c("oracle", "postgresql", "redshift", "sql server", "pdw", "netezza",
    "bigquery", "sqlite"),
  user,
  password,
  server,
  port = NULL,
  schema = NULL
)
}
\arguments{
\item{dataframe}{dataframe from SQL queries}

\item{time_to_rx_col}{numeric variable indicating time to treatment}

\item{cohort_cols}{grouping variables of length 1 or greater}
}
\description{
Plot a histogram of the time to treatment
}
